requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
requests
python-dotenv